load-tests/create/requirements.txt
locust
pika
aio-pika
anyio
python-dotenv
requests